requirements.txt
azure-functions
Flask
openai>=1.0.0
anthropic>=0.20.0
boto3>=1.34.0
